# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrueNewsScanner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /check:
    post:
      operationId: submitCheck
      tags: [check]
      summary: Submit content for checking
      description: Routes input to the appropriate pipeline (URL, phone, message, news)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckRequest"
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /check/{id}:
    get:
      operationId: getCheckById
      tags: [check]
      summary: Get a check result by ID
      description: Retrieve a specific past check result by its UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checks/search:
    get:
      operationId: searchChecks
      tags: [check]
      summary: Search prior check results
      description: Look up previously checked inputs by partial match
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search query (partial match on input)
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/CheckType"
          description: Filter by check type
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckSearchResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checks/top-threats:
    get:
      operationId: getTopThreats
      tags: [check]
      summary: Get top confirmed threats
      description: Returns the highest-risk scam and suspicious entries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/CheckType"
      responses:
        "200":
          description: Top threat list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopThreatList"
  /feedback:
    post:
      operationId: submitFeedback
      tags: [report]
      summary: Submit feedback on a check result
      description: Rate a check result as helpful, misleading, or inaccurate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Check not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feedback/{checkId}/summary:
    get:
      operationId: getFeedbackSummary
      tags: [report]
      summary: Get feedback summary for a check
      description: Returns aggregated feedback ratings for a specific check result
      parameters:
        - name: checkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Feedback summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSummary"
        "404":
          description: Check not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    CheckType:
      type: string
      enum: [url, phone, message, news]
    Verdict:
      type: string
      enum:
        - safe
        - insufficient
        - suspicious
        - misleading
        - scam
    CheckRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/CheckType"
        input:
          type: string
          description: The URL, phone number, message text, or news text to check
      required:
        - type
        - input
    Signal:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
          description: Bulgarian label for the signal
        description:
          type: string
          description: Bulgarian description of what was found
        weight:
          type: number
          description: Contribution to risk score (-1 to 1)
        isRisk:
          type: boolean
          description: True if this signal increases risk
      required:
        - id
        - label
        - description
        - weight
        - isRisk
    Evidence:
      type: object
      properties:
        source:
          type: string
          description: Name of the data source
        finding:
          type: string
          description: Bulgarian description of what was found
        url:
          type: string
          nullable: true
          description: Optional link to source
      required:
        - source
        - finding
    NextStep:
      type: object
      properties:
        action:
          type: string
          description: Bulgarian action label
        description:
          type: string
          description: Bulgarian description
        priority:
          type: string
          enum: [high, medium, low]
      required:
        - action
        - description
        - priority
    CheckResult:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/CheckType"
        input:
          type: string
        verdict:
          $ref: "#/components/schemas/Verdict"
        verdictLabel:
          type: string
          description: Bulgarian verdict label
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 0=safe, 100=definite scam
        confidence:
          type: integer
          minimum: 0
          maximum: 100
          description: Confidence in the assessment
        summary:
          type: string
          description: Bulgarian summary of the assessment
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        nextSteps:
          type: array
          items:
            $ref: "#/components/schemas/NextStep"
        checkedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - input
        - verdict
        - verdictLabel
        - riskScore
        - confidence
        - summary
        - signals
        - evidence
        - nextSteps
        - checkedAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    CheckSummary:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/CheckType"
        input:
          type: string
        verdict:
          $ref: "#/components/schemas/Verdict"
        verdictLabel:
          type: string
        riskScore:
          type: integer
        checkedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - input
        - verdict
        - verdictLabel
        - riskScore
        - checkedAt
    CheckSearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/CheckSummary"
        count:
          type: integer
      required:
        - results
        - count
    ThreatEntry:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/CheckType"
        input:
          type: string
        verdict:
          $ref: "#/components/schemas/Verdict"
        verdictLabel:
          type: string
        riskScore:
          type: integer
        confidence:
          type: integer
        summary:
          type: string
        checkedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - input
        - verdict
        - verdictLabel
        - riskScore
        - confidence
        - summary
        - checkedAt
    TopThreatList:
      type: object
      properties:
        threats:
          type: array
          items:
            $ref: "#/components/schemas/ThreatEntry"
        count:
          type: integer
      required:
        - threats
        - count
    FeedbackRating:
      type: string
      enum: [helpful, misleading, inaccurate]
    FeedbackRequest:
      type: object
      properties:
        checkId:
          type: string
        rating:
          $ref: "#/components/schemas/FeedbackRating"
        comment:
          type: string
          nullable: true
      required:
        - checkId
        - rating
    FeedbackResponse:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - message
        - createdAt
    FeedbackSummary:
      type: object
      properties:
        checkId:
          type: string
        total:
          type: integer
        ratings:
          type: object
          properties:
            helpful:
              type: integer
            misleading:
              type: integer
            inaccurate:
              type: integer
          required:
            - helpful
            - misleading
            - inaccurate
      required:
        - checkId
        - total
        - ratings
...