# lib/api-zod/src/generated/types/checkSearchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrueNewsScanner API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckSummary } from "./checkSummary";

export interface CheckSearchResult {
  results: CheckSummary[];
  count: number;
}