# lib/api-zod/src/generated/types/feedbackRating.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrueNewsScanner API
 * OpenAPI spec version: 0.1.0
 */

export type FeedbackRating =
  (typeof FeedbackRating)[keyof typeof FeedbackRating];

export const FeedbackRating = {
  helpful: "helpful",
  misleading: "misleading",
  inaccurate: "inaccurate",
} as const;