# lib/api-zod/src/generated/types/feedbackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrueNewsScanner API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackRating } from "./feedbackRating";

export interface FeedbackRequest {
  checkId: string;
  rating: FeedbackRating;
  comment?: string | null;
}