# lib/api-zod/src/generated/types/feedbackResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrueNewsScanner API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackResponse {
  id: string;
  message: string;
  createdAt: Date;
}