# lib/api-zod/src/generated/types/threatEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrueNewsScanner API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckType } from "./checkType";
import type { Verdict } from "./verdict";

export interface ThreatEntry {
  id: string;
  type: CheckType;
  input: string;
  verdict: Verdict;
  verdictLabel: string;
  riskScore: number;
  confidence: number;
  summary: string;
  checkedAt: Date;
}