# lib/api-zod/src/generated/types/topThreatList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrueNewsScanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatEntry } from "./threatEntry";

export interface TopThreatList {
  threats: ThreatEntry[];
  count: number;
}